refactor: WritingSubmit 파일 구조 재정리(#47) - WritingSubmit 폴더 안에 WritingSubmit.jsx, WritingSubmit.css 옮김 - hooks 폴더에 useSubmittedWriting.js 옮김
